refactor: rename openapi params schema
