+++ src/openapi/schemas/id-params-schema.ts
